GPT_email_generator/requirements.txt
streamlit
openai
google-generativeai
requests
textblob
gspread
oauth2client
praw
python-dotenv
